# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pokemon Anil API",
    "version": "2.0.1",
    "description": "API propria do Pokemon Anil/PT-BR. Dados locais Azul 4.0.6 (interna 4.0.3). Sem consultas externas nos catalogos de jogo."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/teams": {
      "get": {
        "summary": "GET /teams",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "parameters": [
          {
            "name": "modo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ]
      },
      "post": {
        "summary": "POST /teams",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created, where applicable"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "parameters": [
          {
            "name": "modo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "security": [
          {
            "writeKey": []
          }
        ],
        "description": "Remote writes require API_WRITE_KEY. Local writes are allowed without a configured key; once configured it is required locally too."
      }
    },
    "/teams/{id}": {
      "get": {
        "summary": "GET /teams/{id}",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "put": {
        "summary": "PUT /teams/{id}",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created, where applicable"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "security": [
          {
            "writeKey": []
          }
        ],
        "description": "Remote writes require API_WRITE_KEY. Local writes are allowed without a configured key; once configured it is required locally too."
      },
      "delete": {
        "summary": "DELETE /teams/{id}",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "writeKey": []
          }
        ],
        "description": "Remote writes require API_WRITE_KEY. Local writes are allowed without a configured key; once configured it is required locally too."
      }
    },
    "/teams/{id}/stats": {
      "get": {
        "summary": "GET /teams/{id}/stats",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/teams/{id}/analysis": {
      "get": {
        "summary": "GET /teams/{id}/analysis",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/teams/{id}/members": {
      "post": {
        "summary": "POST /teams/{id}/members",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created, where applicable"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "security": [
          {
            "writeKey": []
          }
        ],
        "description": "Remote writes require API_WRITE_KEY. Local writes are allowed without a configured key; once configured it is required locally too."
      }
    },
    "/teams/{id}/members/{slot}": {
      "delete": {
        "summary": "DELETE /teams/{id}/members/{slot}",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slot",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "writeKey": []
          }
        ],
        "description": "Remote writes require API_WRITE_KEY. Local writes are allowed without a configured key; once configured it is required locally too."
      }
    },
    "/teams/{id}/clone": {
      "post": {
        "summary": "POST /teams/{id}/clone",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created, where applicable"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "External service failure"
          },
          "504": {
            "description": "External service timeout"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "security": [
          {
            "writeKey": []
          }
        ],
        "description": "Remote writes require API_WRITE_KEY. Local writes are allowed without a configured key; once configured it is required locally too."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "writeKey": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}